requirements.txt
aiohttp
fastapi
pillow
pydantic
zxing-cpp
